backend/monitorsol/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
openai-whisper @ git+https://github.com/openai/whisper.git
ffmpeg-python==0.2.0
aiofiles==23.2.1
aiosqlite==0.20.0
python-multipart==0.0.9